added sec:authorize for leftMenu

+++ src/main/java/kg/inaiproject/booking/entities/BonusCard.java
@@ -20,7 +20,7 @@ public class BonusCard {
     Long id;
 
     @Column(name = "balance")
-    BigDecimal balance;
+    Double balance;
 
     @OneToOne
     User user;
